% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{aggregate_param}
\alias{aggregate_param}
\title{Transform Parameter Function into Values by Age Bin}
\usage{
aggregate_param(f_age, target_bins, population, ...)
}
\arguments{
\item{f_age}{a function, \code{f(a, ...)} which takes age \code{a} as the first argument
and optionally any other \code{...}. \code{f} should be vectorized in \code{a}}

\item{target_bins}{a vector, the cutoffs for each age bin. For example, a vector
\code{ages <- c(0, 5, 20, 100)} would correspond to intervals of \verb{[0, 5)}, \verb{[5, 20)},
and \verb{[20, 100]}. Note the first value is a start (rather than cutoff) and
that upper limits are exclusive, \emph{except} for the final value.}

\item{population}{a data.frame, the population by age.}
}
\value{
a vector,
}
\description{
\code{aggregate_param} integrates a functional relationship between some model
feature and age, for example case fatality ratio, into static parameter
values for use in a lower-resolution age-binned compartmental model.
}
