% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{disaggregate_value}
\alias{disaggregate_value}
\title{Disaggregate a Compartmental Model Age-Based Outcome}
\usage{
disaggregate_value(value, f_age, target_bins, population, ...)
}
\arguments{
\item{value}{a numeric vector, the model output by age group}

\item{f_age}{the relative rate of the subject outcome, by high
resolution age. Note, this should be the same \code{f_age} as in \code{\link[=aggregate_param]{aggregate_param()}}
for parameter / outcome pairs - for example, when the outcome is death and
the parameter function is the age-specific case fatality ratio (which would be
aggregated to give single lower age resolution compartmental rates).}

\item{target_bins}{a vector, the cutoffs for each age bin. For example, a vector
\code{ages <- c(0, 5, 20, 100)} would correspond to intervals of \verb{[0, 5)}, \verb{[5, 20)},
and \verb{[20, 100]}. Note the first value is a start (rather than cutoff) and
that upper limits are exclusive, \emph{except} for the final value.}

\item{population}{a data.frame, the population by age.}
}
\description{
\code{disaggregate_value} takes a low-age resolution outcome, for example deaths,
and proportionally distributes that outcome into a higher age resolution for
use in subsequent analyses like years-life-lost style calculations.
}
