% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{distill_summary}
\alias{distill_summary}
\title{Convenient Distillation Comparison Summary}
\usage{
distill_summary(model_outcomes_dt, density_dt, mapping_dt)
}
\arguments{
\item{model_outcomes_dt}{a data.table (or convertable to such) with columns
\code{from} and \code{value}}

\item{model_upper}{the upper limit of the last partition}
}
\description{
Calculates the outcomes for various distillation assumptions.
}
