% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{make_partition}
\alias{make_partition}
\title{Create a Least-Common-Interval Partition}
\usage{
make_partition(...)
}
\arguments{
\item{...}{any number of numeric vectors}
}
\value{
a sorted numeric vector with unique values
}
\description{
Internal utility method for creating partitions, possibly from multiple
distinct partitions. Validates inputs and potential creates open ends.
}
\examples{
paramix:::make_partition(1:5)
paramix:::make_partition(5:1, 4, 6)

}
