% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{parameter_summary}
\alias{parameter_summary}
\title{Convenient Parameter Comparison Summary}
\usage{
parameter_summary(
  f_param,
  densities,
  model_partition,
  ...,
  resolution = diff(range(model_partition)) + 1L
)
}
\arguments{
\item{resolution}{the number of points to calculate for the underlying
\code{f_param} line}
}
\description{
Convenient Parameter Comparison Summary
}
