% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramix-package.R
\docType{package}
\name{paramix-package}
\alias{paramix}
\alias{paramix-package}
\title{paramix: Age-Based Aggregation and Disaggregation for Analysis of Compartmental Models}
\description{
The `paramix` package provides a convenient framework for aggregating and disaggregating continuously varying parameters (for example, case fatality ratio, with age) for proper parametrization of lower-resolution compartmental models (for example, with broad age categories) and subsequent upscaling of model outputs at high resolution (for example, as needed when calculating age-sensitive measures like years-life-lost).
}
\author{
\strong{Maintainer}: Carl Pearson \email{carl.ab.pearson@gmail.com} (\href{https://orcid.org/0000-0003-0701-7860}{ORCID})

Authors:
\itemize{
  \item Simon Proctor \email{simon.procter@lshtm.ac.uk} (\href{https://orcid.org/0000-0002-0380-1503}{ORCID})
  \item Lucy Goodfellow \email{lucy.goodfellow@lshtm.ac.uk} (\href{https://orcid.org/0009-0004-0434-5863}{ORCID})
}

}
\keyword{internal}
