% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{alembic}
\alias{alembic}
\title{Create the Blending and Distilling Object}
\usage{
alembic(
  f_param,
  f_dense,
  model_partition,
  new_partition,
  pars_interp_opts = list(fun = stats::splinefun, method = "natural"),
  dens_interp_opts = list(fun = stats::approxfun, method = "constant", yleft = 0, yright
    = 0)
)
}
\arguments{
\item{f_param}{a function, \code{f(x)} which transforms the feature (e.g. age),
and yields the parameter value. Alternatively, a \code{data.frame} where the first
column is the feature (x) and the second is the parameter (y); see
\code{\link[=xy.coords]{xy.coords()}} for details. If the latter, combined with \code{pars_interp_opts},
and defaulting to spline interpolation.}

\item{f_dense}{like \code{f_param}, either a density function (though it does
not have to integrate to 1 like a pdf) or a \code{data.frame} of values. If the
latter, combined with \code{dens_interp_opts} and defaulting to constant density
from each x to the next.}

\item{model_partition}{a numeric vector of cut points, which define the
partitioning that will be used in the model}

\item{new_partition}{the partition of the underlying feature}

\item{pars_interp_opts}{a list, minimally with an element \code{fun},
corresponding to an interpolation function. Defaults to \code{\link[=splinefun]{splinefun()}}
"natural" interpolation}

\item{dens_interp_opts}{ibid, but for density. Defaults to \code{\link[=approxfun]{approxfun()}}
"constant" interpolation}
}
\value{
a \code{data.frame} which maps fractions of the original model partitions
to the desired partitions, according to underlying relative outcome rates and
densities
}
\description{
Create the Blending and Distilling Object
}
\examples{
ifr_levin <- function(age_in_years) {
  scaled <- exp(-7.56 + 0.121 * age_in_years)
  scaled / (100 + scaled)
}
age_limits <- c(seq(0, 69, by = 5), 70, 80, 100)
age_pyramid <- data.frame(
  from = 0:100, weight = ifelse(0:100 < 65, 1, .99^(0:100-64))
) # flat age distribution, then 1\% annual deaths
ifr_alembic <- alembic(ifr_levin, age_pyramid, age_limits, 0:100)

}
