% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{distill}
\alias{distill}
\title{Re-Map Compartmental Model Outcomes}
\usage{
distill(outcomes, mapping)
}
\arguments{
\item{outcomes}{a long-format \code{data.frame} with columns \code{model_part} and
\code{value}, optionally others that will be preserved}

\item{mapping}{see \code{\link[=alembic]{alembic()}}: a \code{data.frame} in accordance with the return
of that function.}
}
\value{
a \code{data.frame} mirroring \code{outcomes}, but with
}
\description{
\code{distill} takes a low-age resolution outcome, for example deaths,
and proportionally distributes that outcome into a higher age resolution for
use in subsequent analyses like years-life-lost style calculations.
}
