% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{distill}
\alias{distill}
\title{Distill Outcomes}
\usage{
distill(outcomes, alembic_dt)
}
\arguments{
\item{outcomes}{a long-format \code{data.frame} with columns \code{model_from} and
\code{value}, optionally others that will be preserved}

\item{alembic_dt}{an \code{\link[=alembic]{alembic()}} return value}
}
\value{
a \code{data.frame} mirroring \code{outcomes}, but with
}
\description{
\code{distill} takes a low-age resolution outcome, for example deaths,
and proportionally distributes that outcome into a higher age resolution for
use in subsequent analyses like years-life-lost style calculations.
}
